elibrary_project/requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
nltk>=3.8
gensim>=4.3.0
scikit-learn>=1.3.0